refactor: loaders.ts file split into separate files
